refactor(RHIDP-14129): bundle intent-based CLI commands into one directory
Move catalog/api/search/docs/template/backstage-passthrough and their
supporting client/format/intent-errors/helpers modules into
src/commands/intent-based-actions/, mirroring the existing
export-dynamic-plugin/ and package-dynamic-plugins/ layout, with a
single registerIntentCommands() entry point.
